i fucking hate timelines
... spent way to many hours getting the randomisation to work as i wanted because UE is incapable to do what i want it to.

wanted a "go from x to 30 to 0 to 30 where x is random number between 0-30"
ie, 5-30-0 or 22-30-0. ue cant do that without messing with the timelines in ways i dont know how, so i changed it to have a random bool that sets it to 2-5 * 30 giving the coins a bounce between 60 and 150 making them apear to be desynced,,, not what i wanted but f my life right now.